feat(error): update translations and move error translations into own file
