+++ src/routes/$lang/Error.tsx
@@ -1,17 +1,14 @@
-import { LinksFunction, LoaderFunction } from '@remix-run/node';
+import { LoaderFunction } from '@remix-run/node';
 import { useLoaderData } from '@remix-run/react';
 import { useTranslation } from 'react-i18next';
 import { HttpStatusCode } from '~/models/http/statusCodes';
 
-export const handle = {
-  i18n: 'common',
-};
+export const handle = { i18n: 'error' };
 
 export const loader: LoaderFunction = ({
   request,
 }): {
   statusCode: number;
-  url: string;
 } => {
   const queryParameters = new URLSearchParams(request.url.split('?')[1]);
   const queryHttpCode = Number(queryParameters.get('code'));
@@ -19,14 +16,13 @@ export const loader: LoaderFunction = ({
     ? queryHttpCode
     : HttpStatusCode.NotFound;
 
-  return { statusCode, url: request.url };
+  return { statusCode };
 };
 
 const CatchAll = () => {
   const { t } = useTranslation('common');
-  const { statusCode, url } = useLoaderData<{
+  const { statusCode } = useLoaderData<{
     statusCode: HttpStatusCode;
-    url: string;
   }>();
 
   return <div>{t(`error:code_${HttpStatusCode[statusCode]}`)}</div>;
